wip: local unsafe plugins, refactor builtin
