+++ web/src/beta/lib/core/Crust/Widgets/Widget/builtin/Navigator/UI/index.tsx
@@ -4,7 +4,7 @@ import Icon from "@reearth/beta/components/Icon";
 import { useT } from "@reearth/services/i18n";
 import { styled } from "@reearth/services/theme";
 
-import type { Theme } from "../../types";
+import type { Theme } from "../../../types";
 
 import { useNavigator } from "./hooks";
