+++ web/src/beta/lib/core/Crust/Widgets/Widget/builtin/Navigator/hooks.ts
@@ -1,7 +1,7 @@
 import { useState, useEffect, useCallback, useRef } from "react";
 
-import type { Camera, FlyToDestination, Widget } from "../types";
-import { useVisible } from "../useVisible";
+import type { Camera, FlyToDestination, Widget } from "../../types";
+import { useVisible } from "../../useVisible";
 
 import { degreeToRadian, radianToDegree } from "./UI";
