+++ web/src/beta/lib/core/Crust/Widgets/Widget/builtin/Storytelling/index.stories.tsx
@@ -1,6 +1,6 @@
 import { Meta, Story } from "@storybook/react";
 
-import { contextEvents } from "../storybook";
+import { contextEvents } from "../../storybook";
 
 import Component, { Props } from ".";
