+++ web/src/beta/lib/unsafeBuiltinPlugins/.gitignore
@@ -1,0 +1,3 @@
+*
+!.gitignore
+!README.md
