+++ web/src/beta/lib/unsafeBuiltinPlugins/README.md
@@ -1,0 +1,5 @@
+## This folder is for any personal plugins you want to write but treat as builtin, bypassing the sandboxing that happens to regularly installed plugins
+
+# USE AT OWN RISK
+
+TODO: fill in details of how to setup
